refactor(cz-git): adjust project dir
