+++ package.json
@@ -48,6 +48,7 @@
   "devDependencies": {
     "@commitlint/cli": "^16.2.1",
     "@commitlint/config-conventional": "^16.2.1",
+    "@types/inquirer": "^7.3.3",
     "@types/jest": "^27.4.1",
     "@types/node": "^17.0.23",
     "@types/rimraf": "^3.0.2",
