+++ packages/@cz-git/plugin-inquirer/src/index.ts
@@ -1,1 +1,1 @@
-console.log("Hello, world!");
+export * from "./search";
