+++ packages/@cz-git/plugin-inquirer/src/search/index.ts
@@ -1,0 +1,1 @@
+export const a = 1;
