+++ packages/@cz-git/plugin-inquirer/src/shared/index.ts
@@ -1,0 +1,1 @@
+export * from "./types";
