+++ packages/@cz-git/plugin-inquirer/src/shared/types/index.ts
@@ -1,0 +1,1 @@
+export type AA = string;
